even more refactoring into user component
